+++ quesma/quesma/test_utils.go
@@ -1,7 +1,7 @@
 // Copyright Quesma, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-package main
+package quesma
 
 import (
 	"bytes"
